Cli Improvements 1
